Extract and generalize the LogDistributor into a ZmqDistributor.
