+++ catkit2/testbed/__init__.py
@@ -8,6 +8,7 @@ __all__ = [
     'Experiment',
     'TraceWriter',
     'TracingProxy',
+    'ZmqDistributor',
 ]
 
 from .testbed import *
@@ -15,5 +16,6 @@ from .experiment import *
 from .service import *
 from .logging import *
 from .tracing import *
+from .distributor import *
 from .testbed_proxy import *
 from .service_proxy import *
