+++ catkit2/testbed/logging.py
@@ -3,7 +3,6 @@ import threading
 import zmq
 import json
 import contextlib
-import traceback
 from colorama import Fore, Back, Style
 
 from ..catkit_bindings import submit_log_entry, Severity
@@ -26,79 +25,6 @@ class CatkitLogHandler(logging.StreamHandler):
         severity = getattr(Severity, record.levelname)
 
         submit_log_entry(filename, line, function, severity, message)
-
-class LogDistributor:
-    '''Collects log messages on a port and re-publish them on another.
-
-    This operates on a separate thread after it is started.
-
-    Parameters
-    ----------
-    context : zmq.Context
-        A previously-created ZMQ context. All sockets will be created on this context.
-    input_port : integer
-        The port number for the incoming log messages.
-    output_port : integer
-        The port number for the outgoing log messages.
-    '''
-    def __init__(self, context, input_port, output_port):
-        self.context = context
-        self.input_port = input_port
-        self.output_port = output_port
-
-        self.shutdown_flag = threading.Event()
-        self.thread = None
-
-    def start(self):
-        '''Start the proxy thread.
-        '''
-        self.thread = threading.Thread(target=self.forwarder)
-        self.thread.start()
-
-    def stop(self):
-        '''Stop the proxy thread.
-
-        This function waits until the thread is actually stopped.
-        '''
-        self.shutdown_flag.set()
-
-        if self.thread:
-            self.thread.join()
-
-    def forwarder(self):
-        '''Create sockets and republish all received log messages.
-
-        .. note::
-            This function should not be called directly. Use
-            :func:`~catkit2.testbed.LoggingProxy.start` to start the proxy.
-        '''
-        collector = self.context.socket(zmq.PULL)
-        collector.RCVTIMEO = 50
-        collector.bind(f'tcp://*:{self.input_port}')
-
-        publicist = self.context.socket(zmq.PUB)
-        publicist.bind(f'tcp://*:{self.output_port}')
-
-        while not self.shutdown_flag.is_set():
-            try:
-                try:
-                    log_message = collector.recv_multipart()
-                    publicist.send_multipart(log_message)
-                except zmq.ZMQError as e:
-                    if e.errno == zmq.EAGAIN:
-                        # Timed out.
-                        continue
-                    else:
-                        raise RuntimeError('Error during receive') from e
-
-                #log_message = log_message[0].decode('utf-8')
-                #log_message = json.loads(log_message)
-
-                #print(f'[{log_message["service_id"]}] {log_message["message"]}')
-            except Exception:
-                # Something went wrong during handling of the log message.
-                # Let's ignore this error, but still print the exception.
-                print(traceback.format_exc())
 
 class LogObserver:
     def __init__(self, host, port):
